+++ lib/src/callback_dispatcher.dart
@@ -17,10 +17,10 @@ void callbackDispatcher() {
       final handle = CallbackHandle.fromRawHandle(args[0] as int);
       final id = args[1] as String;
       final status = args[2] as int;
-      final progress = args[3] as int;
+      final progress = args[3] as double;
 
       final callback = PluginUtilities.getCallbackFromHandle(handle) as void
-          Function(String id, int status, int progress)?;
+          Function(String id, int status, double progress)?;
 
       if (callback == null) {
         // The callback wasn't registered. Ignore.
